Add Top Floods Group
Adds a Group to record all Top Flood Lights. Center Stage uses the backdrop for Green to keep everything uniform.
(Not a bug fix, but I've seen quite a few requests for it.)
Does not include LIVE flood in this group. During testing it often got in the way. You can still add the LIVE flood via the original Lights CS Group.